requirements.txt
flask==2.0.1
requests==2.26.0
python-dotenv==0.19.0
praw==7.5.0
google-api-python-client==2.65.0
pytrends==4.8.0
python-dateutil==2.8.2
tweepy==4.10.0
beautifulsoup4==4.9.3 